code migration from functions.cpp to interrupts.cpp (only 1 file)
